(feat) moved index.js from top level of public to components dir
