+++ src/public/AppEntry.js
@@ -1,5 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './index.js';
+import App from './components/index.js';
 
 ReactDOM.render(<App />, document.getElementById('app'));
