+++ src/public/components/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Routes from './Routes';
+import Routes from '../Routes';
 import autoBind from 'react-autobind';
-import Navigation from './components/Navigation';
-import { get } from './components/AxiosRoutes';
+import Navigation from './Navigation';
+import { get } from './AxiosRoutes';
 import { BrowserRouter as Router, Route, Link, Redirect } from 'react-router-dom';
 
 export default class App extends React.Component {
